Init for v1.0.0

+++ encoder/encoder.go
@@ -40,5 +40,5 @@ func ToBase62(u uint64) string {
 // padding zeros
 func ToBase62WithPaddingZeros(u uint64, padding int64) string {
 	formatter := "%+0" + strconv.FormatInt(padding, 10) + "s"
-	return fmt.Sprintf(formatter, ToBase62(uint64(u)))
+	return fmt.Sprintf(formatter, ToBase62(u))
 }
